libs: snapshot: extract image and EXIF helpers from store
Move buildExif, embed helpers, filename, and thumbnail generation into the
snapshot lib so the Pinia store only orchestrates capture, storage, and
downloads, and consumers can import image helpers without the store.
