+++ src/stores/snapshot.ts
@@ -1,20 +1,17 @@
 import { useThrottleFn } from '@vueuse/core'
 import { BlobReader, BlobWriter, ZipWriter } from '@zip.js/zip.js'
-import { format } from 'date-fns'
 import saveAs from 'file-saver'
-import piexif from 'piexifjs'
 import { defineStore } from 'pinia'
 
 import { useInteractionDialog } from '@/composables/interactionDialog'
 import { useBlueOsStorage } from '@/composables/settingsSyncer'
-import { app_version } from '@/libs/cosmos'
 import { availableCockpitActions, registerActionCallback } from '@/libs/joystick/protocols/cockpit-actions'
-import { formatExifDateTime, formatExifGpsDateStamp, toExifGpsTimeStamp } from '@/libs/snapshot'
-import { isElectron, sanitizeFilenameComponent } from '@/libs/utils'
+import { buildExif, createThumbnail, maybeEmbedExif, snapshotFilename } from '@/libs/snapshot'
+import { isElectron } from '@/libs/utils'
 import { snapshotStorage, snapshotThumbStorage } from '@/libs/videoStorage'
 import { useMissionStore } from '@/stores/mission'
 import { StorageDB } from '@/types/general'
-import { EIXFType, SnapshotExif, SnapshotFileDescriptor, SnapshotResult } from '@/types/snapshot'
+import { SnapshotFileDescriptor, SnapshotResult } from '@/types/snapshot'
 import { DownloadProgressCallback, FileDescriptor } from '@/types/video'
 
 import { useMainVehicleStore } from './mainVehicle'
@@ -27,86 +24,6 @@ export const useSnapshotStore = defineStore('snapshot', () => {
   const { showDialog } = useInteractionDialog()
 
   const zipMultipleFiles = useBlueOsStorage('cockpit-zip-multiple-video-files', false)
-
-  const blobToDataURL = (blob: Blob): Promise<string> =>
-    new Promise((res) => {
-      const r = new FileReader()
-      r.onload = () => res(r.result as string)
-      r.readAsDataURL(blob)
-    })
-
-  const dataURLToBlob = (dataURL: string): Blob => {
-    const [meta, b64] = dataURL.split(',')
-    const bin = atob(b64)
-    const arr = new Uint8Array(bin.length)
-    for (let i = 0; i < bin.length; i++) arr[i] = bin.charCodeAt(i)
-    const mime = /data:(.*);base64/.exec(meta)?.[1] ?? 'image/jpeg'
-    return new Blob([arr], { type: mime })
-  }
-
-  const toDMS = (deg: number): [number, number][] => {
-    const d = Math.floor(Math.abs(deg))
-    const mFloat = (Math.abs(deg) - d) * 60
-    const m = Math.floor(mFloat)
-    const s = Math.round((mFloat - m) * 6000)
-    return [
-      [d, 1],
-      [m, 1],
-      [s, 100],
-    ]
-  }
-
-  const buildExif = (opts: EIXFType): SnapshotExif => {
-    const { latitude, longitude, capturedAt, yaw, pitch, roll, width, height } = opts
-    const dateTime = formatExifDateTime(capturedAt)
-
-    const jsonComment = {
-      vehicle_attitude: {
-        yaw: yaw ?? 0,
-        pitch: pitch ?? 0,
-        roll: roll ?? 0,
-      },
-    }
-
-    return {
-      '0th': {
-        [piexif.ImageIFD.Software]: `Cockpit ${app_version.version} - Blue Robotics`,
-        [piexif.ImageIFD.DateTime]: dateTime,
-      },
-      'Exif': {
-        [piexif.ExifIFD.UserComment]: JSON.stringify(jsonComment),
-        [piexif.ExifIFD.PixelXDimension]: width,
-        [piexif.ExifIFD.PixelYDimension]: height,
-        [piexif.ExifIFD.DateTimeOriginal]: dateTime,
-        [piexif.ExifIFD.DateTimeDigitized]: dateTime,
-      },
-      'GPS': {
-        [piexif.GPSIFD.GPSLatitudeRef]: latitude >= 0 ? 'N' : 'S',
-        [piexif.GPSIFD.GPSLongitudeRef]: longitude >= 0 ? 'E' : 'W',
-        [piexif.GPSIFD.GPSLatitude]: toDMS(Math.abs(latitude)),
-        [piexif.GPSIFD.GPSLongitude]: toDMS(Math.abs(longitude)),
-        [piexif.GPSIFD.GPSDateStamp]: formatExifGpsDateStamp(capturedAt),
-        [piexif.GPSIFD.GPSTimeStamp]: toExifGpsTimeStamp(capturedAt),
-      },
-    }
-  }
-
-  const embedExif = async (blob: Blob, exifObj: piexif.Ifd): Promise<Blob> => {
-    const dataURL = await blobToDataURL(blob)
-    const exifStr = piexif.dump(exifObj)
-    const updated = piexif.insert(exifStr, dataURL)
-    return dataURLToBlob(updated)
-  }
-
-  const maybeEmbedExif = async (b: Blob, exif: piexif.Ifd): Promise<Blob> => {
-    if (b.type !== 'image/jpeg') return b
-    try {
-      return await embedExif(b, exif)
-    } catch (err) {
-      console.error('EXIF embed failed – storing image without metadata', err)
-      return b
-    }
-  }
 
   const captureStreamFrame = async (streamName: string): Promise<Blob> => {
     if (!streamName) {
@@ -183,50 +100,6 @@ export const useSnapshotStore = defineStore('snapshot', () => {
     return new Blob([jpegBuffer], { type: 'image/jpeg' })
   }
 
-  const snapshotFilename = (streamName: string, missionName = 'Cockpit'): string => {
-    // Sanitize because the `O` timezone token emits a real colon for non-integer UTC offsets (e.g. `GMT+5:30`), which is illegal on Windows and yields a 0KB file.
-    const timeString = sanitizeFilenameComponent(format(new Date(), 'LLL dd, yyyy - HH꞉mm꞉ss.SSS O'))
-    const safeMissionName = sanitizeFilenameComponent(missionName) || 'Cockpit'
-    const safeStreamName = sanitizeFilenameComponent(streamName) || 'workspace'
-    return `${safeMissionName} (${timeString}) #${safeStreamName}.jpeg`
-  }
-
-  const createThumbnail = (blob: Blob, width: number, height: number): Promise<Blob> => {
-    const img = document.createElement('img')
-    const objectUrl = URL.createObjectURL(blob)
-    img.src = objectUrl
-    img.width = width
-    img.height = height
-
-    const canvas = document.createElement('canvas')
-    canvas.width = width
-    canvas.height = height
-    const ctx = canvas.getContext('2d')
-    if (!ctx) throw new Error('Could not get 2D context')
-
-    return new Promise<Blob>((resolve, reject) => {
-      img.onload = () => {
-        ctx.drawImage(img, 0, 0, width, height)
-        canvas.toBlob(
-          (thumbnailBlob) => {
-            URL.revokeObjectURL(objectUrl)
-            if (thumbnailBlob) {
-              resolve(thumbnailBlob)
-            } else {
-              reject(new Error('Canvas toBlob failed'))
-            }
-          },
-          'image/jpeg',
-          0.9
-        )
-      }
-      img.onerror = () => {
-        URL.revokeObjectURL(objectUrl)
-        reject(new Error('Image load failed'))
-      }
-    })
-  }
-
   /**
    * Captures snapshots from the given streams and optionally the workspace.
    * @param {string[]} streamNames - Names of the video streams to capture
@@ -258,7 +131,7 @@ export const useSnapshotStore = defineStore('snapshot', () => {
           })
           wsBlob = await maybeEmbedExif(wsBlob, wsExif)
           const thumbBlob = await createThumbnail(wsBlob, 200, 113)
-          const filename = snapshotFilename('workspace', missionName)
+          const filename = snapshotFilename('workspace', missionName, capturedAt)
           const thumbFilename = filename + '-thumb'
           await snapshotStorage.setItem(filename, wsBlob)
           await snapshotThumbStorage.setItem(thumbFilename, thumbBlob)
@@ -278,7 +151,7 @@ export const useSnapshotStore = defineStore('snapshot', () => {
         const stExif = buildExif({ latitude, longitude, capturedAt, yaw, pitch, roll, width, height })
         stBlob = await maybeEmbedExif(stBlob, stExif)
         const internalStreamName = videoStore.internalStreamNameFromExternal(streamName) ?? streamName
-        const filename = snapshotFilename(internalStreamName, missionName)
+        const filename = snapshotFilename(internalStreamName, missionName, capturedAt)
         const thumbFilename = filename + '-thumb'
 
         await snapshotStorage.setItem(filename, stBlob)
@@ -374,9 +247,7 @@ export const useSnapshotStore = defineStore('snapshot', () => {
   return {
     snapshotStorage,
     snapshotThumbStorage,
-    createThumbnail,
     downloadFilesFromSnapshotDB,
-    snapshotFilename,
     takeSnapshot,
     deleteSnapshotFiles,
     zipMultipleFiles,
